Separate swell option for passing SLURM directives
Additional option `-s somefile.yaml` for specifying additional SLURM
directives.

Note that the implementation here adds the SLURM overrides to the
internal `experiment_dict` (and, ultimately, the `experiment.yaml`) with
a bit of error checking to make sure we _only_ use this for SLURM
directives. **This is a bit hacky** --- we should probably either keep this
information completely separate from the `experiment_dict`, or make it a
proper first-class citizen, with Questionary support, etc.