Remove the `advanced` keyword
It's not actually used anywhere in the calling functions; it's just
taking up unnecessary space.

I'm also not convinced that, conceptually, it's a good idea to include
it at all. A better approach to "advanced" configurations is probably to
expose _all_ options to users but provide "advanced" options with
sensible defaults.