Fix GEOS_mksi test
Since https://github.com/GEOS-ESM/GEOS_mksi/pull/2 was merged, GEOS_mksi
main and develop are now in sync. This was causing the GEOS_mksi test
to fail.

This changes the test to point to a known (very old!) version of
GEOS_mksi that will cause an error, so the test actually passes.