Remove hardcoded /tmp from code_tests, and fix GEOS_mksi test (#436)
* Remove hardcoded /tmp from code_tests

See #432

* Fix GEOS_mksi test

Since https://github.com/GEOS-ESM/GEOS_mksi/pull/2 was merged, GEOS_mksi
main and develop are now in sync. This was causing the GEOS_mksi test
to fail.

This changes the test to point to a known (very old!) version of
GEOS_mksi that will cause an error, so the test actually passes.